Create App Product
We created the app product and also it's model and serializers, we also populated it's database. In it's views.py we used only class based views and they inherited the generics class from the djangorestframework. We implemented a full CRUD and if you look at our client.py we also gave it certain scripts to test our api